+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- intégration de l'api Wabap
+
 ## [1.0.7] - 2022-06-15
 
 - moved to NX monorepo